Version Bump - setup.py problems
Given that the Python setup.py does not actually run 'clean' in the
normal sense of the build command (removing all artifacts of previous
builds), version 0.1 is uninstallable.  This build contains no
functional changes whatsoever, in fact the only change at all is the
version number because PyPI has no concept of marking a build as
broken and removing it to free the version number.