Version 0.1.2
This version includes a bugfix for adding devices on behalf of other
users and a new field for hostname.  The hostname is validated against
the rules for RFC compliant hostnames.

The example daemon has also been updated to use a correct path to
Python as well as correct syntax for DHCPd.